requirements.txt
Flask
Flask-SQLAlchemy
requests
Werkzeug
psycopg2-binary
python-telegram-bot==20.5
